Monte Carlo Simulation
What Is a Monte Carlo Simulation?
A Monte Carlo simulation is used to model the probability of different outcomes in a process that cannot easily be predicted due to the intervention of random variables. It is a technique used to understand the impact of risk and uncertainty.

A Monte Carlo simulation is used to tackle a range of problems in many fields, including investing, business, physics, and engineering. It is also referred to as a multiple probability simulation.

KEY TAKEAWAYS
A Monte Carlo simulation is a model used to predict the probability of a variety of outcomes when the potential for random variables is present.
Monte Carlo simulations help to explain the impact of risk and uncertainty in prediction and forecasting models.
A Monte Carlo simulation requires assigning multiple values to an uncertain variable to achieve multiple results and then averaging the results to obtain an estimate.
Monte Carlo simulations assume perfectly efficient markets.